Feature(015): - Added language lines. - Refactor UserCard component. - Created UserActionMenu component. - Created WalletAccordion component.
