+++ src/assets/language/en.tsx
@@ -1,8 +1,32 @@
 export const ENG = {
     home: {
         usercard: {
+            actions: {
+                profile: 'Profile',
+                lockUser: 'Lock user'
+            },
             header: {
                 title: 'Users'
+            },
+            wallet: {
+                btc: {
+                    name: 'Bitcoin wallet'
+                },
+                ethereum: {
+                    name: 'Ethereum wallet'
+                },
+                info: {
+                    founds: 'Founds',
+                    lastTransaction: {
+                        name: 'Last transaction',
+                        hash: 'Hash',
+                        amount: 'Amount',
+                        date: 'Date'
+                    }
+                },
+                actions: {
+                    addFounds: 'Add founds'
+                }
             }
         },
         operations: {
